Fixed NTR mode switch and fixed icon
Fixed NTR mode switch with new mode switch method. Improved icon and
fixed pallette. (seems first palette entry is alpha by default so that
explains why the background became alpha since the color for it was
first in the palette)

Renamed template.c to main.c in arm7 and updated
patch_ndsheader_dsiware.py to use the correct values for 0x1b8

argv stuff should probly be removed. It's why the bootsplash runs twice
for NTR mode. The bootstrap argv file has it run a second copy of the
bootstrap SRL. Not sure why it does this instead of just booting the
target SRL directly ...