Add AP patches and its loading mechanism
As of commit 27b20e08794a42efba4cdf1f3cbe09c5de5943c3 from
TWiLightMenu.

The code is pretty much a direct copy from TWiLight, with the fcopy BSS
a little smaller (we don't need to alloc 32KB of memory for copying,
the files are on average less than 512 bytes. So only 512 bytes are
alloc'd for now.)

The packaging of the apfixes is removed, and built directly into
NitroFS instead. Not much is lost here, as the only reason we were
using that packing was solely because small file sizes meant slower
extraction for Universal Updater and the like.

Ideally, we'd place the config function in loadFromSD. But it keeps
crashing there, but it doesn't crash in load_conf. So just keep it
there for now.